+++ pypgx/api/utils.py
@@ -406,7 +406,7 @@ def compute_copy_number(
     read_depth, control_statistics, samples_without_sv=None
 ):
     """
-    Compute copy number from read depth for the target gene.
+    Compute copy number from read depth for target gene.
 
     The method will convert read depth from target gene to copy number by
     performing intra-sample normalization using summary statistics from